+++ lib/src/picker/cupertino_picker_overlay.dart
@@ -182,9 +182,7 @@ class _CupertinoPickerOverlayState extends State<CupertinoPickerOverlay> {
 
     return PopScope(
       canPop: false,
-      onPopInvokedWithResult: (_, __) {
-        _closeOverlay();
-      },
+      onPopInvokedWithResult: (_, __) => _closeOverlay(),
       child: Stack(
         clipBehavior: Clip.none,
         children: <Widget>[