+++ lib/src/picker/picker.dart
@@ -2,7 +2,6 @@
 // Use of this source code is governed by a MIT-style license that can be
 // found in the LICENSE file.
 
-export '../calendar/actions/actions.dart';
 export 'button/button.dart';
 export 'container/container.dart';
 export 'cupertino_picker_animated_cross_fade.dart';